Added delay between taking pictures, and print RTC time

+++ EPII_CM55M_APP_S/app/ww_projects/allon_sensor_tflm_freertos/hardfault_handler.c
@@ -46,7 +46,9 @@ void HardFault_Handler(void) {
 		}
 		if (SAU->SFSR & SAU_SFSR_SFARVALID_Msk) {
 			/* SFARVALID: SFAR contain valid address that caused secure violation */
-			printf("Address that caused SAU violation is 0x%X.\r\n", SAU->SFAR);
+			// CGP remove compiler warning
+			// printf("Address that caused SAU violation is 0x%X.\r\n", SAU->SFAR);
+			printf("Address that caused SAU violation is 0x%X.\r\n", (unsigned int) SAU->SFAR);
 		}
 	}
 
@@ -64,8 +66,10 @@ void HardFault_Handler(void) {
 		}
 		if (SCB->CFSR & SCB_CFSR_BFARVALID_Msk) {
 			/* BFARVALID: BFAR contain valid address that caused secure violation */
-			printf("Address that caused secure bus violation is 0x%X.\r\n",
-					SCB->BFAR);
+
+			// CGP remove compiler warning
+			//printf("Address that caused secure bus violation is 0x%X.\r\n", SCB->BFAR);
+			printf("Address that caused secure bus violation is 0x%X.\r\n", (unsigned int) SCB->BFAR);
 		}
 	}
 
@@ -84,8 +88,9 @@ void HardFault_Handler(void) {
 		}
 		if (SCB_NS->CFSR & SCB_CFSR_BFARVALID_Msk) {
 			/* BFARVALID: BFAR contain valid address that caused secure violation */
-			printf("Address that caused secure bus violation is 0x%X.\r\n",
-					SCB_NS->BFAR);
+			// CGP remove compiler warning
+			// printf("Address that caused secure bus violation is 0x%X.\r\n", SCB_NS->BFAR);
+			printf("Address that caused secure bus violation is 0x%X.\r\n", (unsigned int) SCB_NS->BFAR);
 		}
 	}
 
@@ -94,9 +99,14 @@ void HardFault_Handler(void) {
     SCB->AIRCR =
         (SCB->AIRCR & ~SCB_AIRCR_VECTKEY_Msk) | (0x05FAUL << SCB_AIRCR_VECTKEY_Pos) | SCB_AIRCR_SYSRESETREQ_Msk;
 #else
-	printf("SCB->CFSR:0x%08x\n", SCB->CFSR);
-	printf("SCB->BFAR:0x%08x\n", SCB->BFAR);
-	printf("SCB->HFSR:0x%08x\n", SCB->HFSR);
+
+	// CGP remove compiler warning
+//	printf("SCB->CFSR:0x%08x\n", SCB->CFSR);
+//	printf("SCB->BFAR:0x%08x\n", SCB->BFAR);
+//	printf("SCB->HFSR:0x%08x\n", SCB->HFSR);
+	printf("SCB->CFSR:0x%08x\n", (unsigned int) SCB->CFSR);
+	printf("SCB->BFAR:0x%08x\n", (unsigned int) SCB->BFAR);
+	printf("SCB->HFSR:0x%08x\n", (unsigned int) SCB->HFSR);
 	for (;;) {
 	}
 #endif
@@ -115,9 +125,13 @@ void MemManage_Handler(void) {
 }
 void BusFault_Handler(void) {
 	printf("\r\nEntering BusFault_Handler interrupt!\r\n");
-	printf("SCB->CFSR:0x%08x\n", SCB->CFSR);
-	printf("SCB->BFAR:0x%08x\n", SCB->BFAR);
-	printf("SCB->HFSR:0x%08x\n", SCB->HFSR);
+	// CGP remove compiler warning
+//	printf("SCB->CFSR:0x%08x\n", SCB->CFSR);
+//	printf("SCB->BFAR:0x%08x\n", SCB->BFAR);
+//	printf("SCB->HFSR:0x%08x\n", SCB->HFSR);
+	printf("SCB->CFSR:0x%08x\n", (unsigned int) SCB->CFSR);
+	printf("SCB->BFAR:0x%08x\n", (unsigned int) SCB->BFAR);
+	printf("SCB->HFSR:0x%08x\n", (unsigned int) SCB->HFSR);
 	for (;;) {
 	}
 }
